Cleaned up the application factory
This moves all of the fugly code from the main application factory into an api.py module
It also moves the index view into views.py; this breaks one of the tests but this will be remedied in the next commit.
